refactored audio init
