+++ src/system.rs
@@ -17,7 +17,7 @@ use stm32h7xx_hal::stm32::TIM2;
 use stm32h7xx_hal::timer::Event;
 use stm32h7xx_hal::timer::Timer;
 
-use crate::audio;
+use crate::audio::Audio;
 use crate::*;
 
 const HSE_CLOCK_MHZ: MegaHertz = MegaHertz(16);
@@ -40,24 +40,6 @@ const PLL2_R_HZ: Hertz = Hertz(PLL2_P_HZ.0 / 4); // No divder given, what's the 
 const PLL3_P_HZ: Hertz = Hertz(AUDIO_SAMPLE_HZ.0 * 257);
 const PLL3_Q_HZ: Hertz = Hertz(PLL3_P_HZ.0 / 4);
 const PLL3_R_HZ: Hertz = Hertz(PLL3_P_HZ.0 / 16);
-
-// Process samples at 1000 Hz
-// With a circular buffer(*2) in stereo (*2)
-pub const BLOCK_SIZE_MAX: usize = 48;
-pub const DMA_BUFFER_SIZE: usize = BLOCK_SIZE_MAX * 2 * 2;
-
-pub type DmaBuffer = [u32; DMA_BUFFER_SIZE];
-
-const SLOTS: u8 = 2;
-const FIRST_BIT_OFFSET: u8 = 0;
-
-const START_OF_DRAM2: u32 = 0x30000000;
-#[link_section = ".sram1_bss"]
-#[no_mangle]
-static mut TX_BUFFER: DmaBuffer = [0; DMA_BUFFER_SIZE];
-#[link_section = ".sram1_bss"]
-#[no_mangle]
-static mut RX_BUFFER: DmaBuffer = [0; DMA_BUFFER_SIZE];
 
 pub struct System {
     pub gpio: crate::gpio::GPIO,
@@ -221,69 +203,19 @@ impl System {
         .into();
 
         info!("Setup up DMA...");
-        const DMA_MEM_SIZE: usize = 32 * 1024;
-        crate::mpu::dma_init(
+        let audio = Audio::init(
+            device.DMA1,
+            ccdr.peripheral.DMA1,
+            device.SAI1,
+            ccdr.peripheral.SAI1,
+            gpioe.pe2,
+            gpioe.pe3,
+            gpioe.pe4,
+            gpioe.pe5,
+            gpioe.pe6,
+            &ccdr.clocks,
             &mut core.MPU,
             &mut core.SCB,
-            START_OF_DRAM2 as *mut u32,
-            DMA_MEM_SIZE,
-        );
-
-        let dma1_streams = dma::dma::StreamsTuple::new(device.DMA1, ccdr.peripheral.DMA1);
-
-        // dma1 stream 0
-        let tx_buffer: &'static mut [u32; DMA_BUFFER_SIZE] = unsafe { &mut TX_BUFFER };
-        let dma_config = dma::dma::DmaConfig::default()
-            .priority(dma::config::Priority::High)
-            .memory_increment(true)
-            .peripheral_increment(false)
-            .circular_buffer(true)
-            .fifo_enable(false);
-        let dma1_str0: dma::Transfer<_, _, dma::MemoryToPeripheral, _, _> = dma::Transfer::init(
-            dma1_streams.0,
-            unsafe { pac::Peripherals::steal().SAI1 },
-            tx_buffer,
-            None,
-            dma_config,
-        );
-
-        // dma1 stream 1
-        let rx_buffer: &'static mut [u32; DMA_BUFFER_SIZE] = unsafe { &mut RX_BUFFER };
-        let dma_config = dma_config
-            .transfer_complete_interrupt(true)
-            .half_transfer_interrupt(true);
-        let dma1_str1: dma::Transfer<_, _, dma::PeripheralToMemory, _, _> = dma::Transfer::init(
-            dma1_streams.1,
-            unsafe { pac::Peripherals::steal().SAI1 },
-            rx_buffer,
-            None,
-            dma_config,
-        );
-
-        info!("Setup up SAI...");
-        let sai1_rec = ccdr.peripheral.SAI1.kernel_clk_mux(SAI1SEL_A::PLL3_P);
-        let master_config = I2SChanConfig::new(I2SDir::Tx).set_frame_sync_active_high(true);
-        let slave_config = I2SChanConfig::new(I2SDir::Rx)
-            .set_sync_type(I2SSync::Internal)
-            .set_frame_sync_active_high(true);
-
-        let pins_a = (
-            gpioe.pe2.into_alternate_af6(),       // MCLK_A
-            gpioe.pe5.into_alternate_af6(),       // SCK_A
-            gpioe.pe4.into_alternate_af6(),       // FS_A
-            gpioe.pe6.into_alternate_af6(),       // SD_A
-            Some(gpioe.pe3.into_alternate_af6()), // SD_B
-        );
-
-        // Hand off to audio module
-        let dev_audio = device.SAI1.i2s_ch_a(
-            pins_a,
-            AUDIO_SAMPLE_HZ,
-            I2SDataSize::BITS_24,
-            sai1_rec,
-            &ccdr.clocks,
-            master_config,
-            Some(slave_config),
         );
 
         // Setup GPIOs
@@ -322,17 +254,6 @@ impl System {
             Some(gpiob.pb14),
             Some(gpiob.pb15),
         );
-
-        let audio;
-        unsafe {
-            audio = audio::Audio::new(
-                dev_audio,
-                dma1_str1,
-                dma1_str0,
-                &mut RX_BUFFER,
-                &mut TX_BUFFER,
-            );
-        }
 
         // Setup cache
         Self::init_cache(&mut core.SCB, &mut core.CPUID);
